@@ -1,17 +1,15 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdk 33
     defaultConfig {
         applicationId "com.wairdell.learnhelp"
         minSdkVersion 21
-        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        targetSdk 31
 
         externalNativeBuild {
             cmake {
@@ -46,12 +44,13 @@ android {
     }
 
     buildFeatures {
-//        compose true//并在应用的 build.gradle 文件中启用 Jetpack Compose
+        compose true//并在应用的 build.gradle 文件中启用 Jetpack Compose
+        viewBinding true
     }
-    /*composeOptions {
-        kotlinCompilerExtensionVersion compose_version
-        kotlinCompilerVersion kotlin_version
-    }*/
+    composeOptions {
+        kotlinCompilerExtensionVersion "1.4.3"
+//        kotlinCompilerVersion kotlin_version
+    }
 
 }
 
@@ -65,7 +64,7 @@ dependencies {
     implementation "androidx.coordinatorlayout:coordinatorlayout:1.1.0"
     implementation "com.google.android.material:material:1.1.0"
     implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.0.0"
-    implementation "androidx.paging:paging-runtime-ktx:2.1.2"
+    implementation "androidx.paging:paging-runtime-ktx:3.1.1"
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test:runner:1.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
@@ -74,9 +73,8 @@ dependencies {
     implementation 'org.kodein.di:kodein-di-framework-android-x:6.4.1'
     implementation 'org.kodein.di:kodein-di-framework-android-core:6.4.1'
 
-    /*implementation "androidx.compose.ui:ui:$compose_version"
-    implementation "androidx.compose.material:material:$compose_version"
-    implementation "androidx.ui:ui-tooling:$compose_version"*/
+    implementation "androidx.compose.ui:ui:1.3.3"
+    implementation "androidx.compose.material:material:1.3.1"
 
 
 }
