@@ -1,18 +1,22 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.6.10'
-    ext.compose_version = '1.0.0-alpha03'
+    ext.kotlin_version = '1.8.10'
+    ext.compose_version = '1.4.3'
     repositories {
-        maven{ url 'https://maven.aliyun.com/repository/google' }
-        google()
+        google() {
+            allowInsecureProtocol = true
+        }
+        mavenCentral() {
+            allowInsecureProtocol = true
+        }
+        /*maven{ url 'https://maven.aliyun.com/repository/google' }
         maven { url "https://jitpack.io" }
         maven{ url'http://maven.aliyun.com/nexus/content/groups/public/' }
-        maven{ url'http://maven.aliyun.com/nexus/content/repositories/jcenter'}
-        //jcenter()
+        maven{ url'http://maven.aliyun.com/nexus/content/repositories/jcenter'}*/
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath 'com.android.tools.build:gradle:7.4.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -21,12 +25,16 @@ buildscript {
 
 allprojects {
     repositories {
-        maven{ url 'https://maven.aliyun.com/repository/google' }
-        google()
+        google() {
+            allowInsecureProtocol = true
+        }
+        mavenCentral() {
+            allowInsecureProtocol = true
+        }
+       /* maven{ url 'https://maven.aliyun.com/repository/google' }
         maven { url "https://jitpack.io" }
         maven{ url'http://maven.aliyun.com/nexus/content/groups/public/' }
-        maven{ url'http://maven.aliyun.com/nexus/content/repositories/jcenter'}
-        //jcenter()
+        maven{ url'http://maven.aliyun.com/nexus/content/repositories/jcenter'}*/
 
     }
 }
