@@ -9,7 +9,7 @@ open class ViewModelFactory : ViewModelProvider.Factory {
 
     }
 
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return modelClass.newInstance()
     }
 }
